% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRACoD.R
\name{convergence_tests}
\alias{convergence_tests}
\title{Perform convergence tests on the p and beta variables}
\usage{
convergence_tests(trace, df_relab)
}
\arguments{
\item{trace}{the output of run_bracod()}

\item{df_relab}{the microbiome relative abundance}
}
\value{
no return value
}
\description{
You may get errors are divergence of some variables after pymc3 samples the posterior.
We are not overly concerned about some of the variables, such as the variance, rather
we are really interested in the inclusion probabilities (p) and contribution coefficients
(beta). The convergence tests that are included here focus on evaluating those two variables.
}
