% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRACoD.R
\name{remove_null}
\alias{remove_null}
\title{Remove NULL values in your OTU and environmental variable}
\usage{
remove_null(df_relab, Y)
}
\arguments{
\item{df_relab}{microbiome relative abundance data in a dataframe}

\item{Y}{values of the environmental variable}
}
\value{
a list containing 1) the relative abundance data and 2) the Y values
}
\description{
This will remove samples that are NULL in the environmental variable, as well as
the corresponding samples in your relative abundance data.
}
