% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRACoD.R
\name{threshold_count_data}
\alias{threshold_count_data}
\title{Threshold your microbiome counts data}
\usage{
threshold_count_data(df_counts, min_counts = 1000, min_ab = 1e-04)
}
\arguments{
\item{df_counts}{A dataframe of OTU counts. Samples are rows and bacteria are columns.}

\item{min_counts}{threshold samples with fewer than this many counts}

\item{min_ab}{threshold bacteria whose average log abundance is below this}
}
\value{
a dataframe of microbiome counts
}
\description{
This function removes samples below a minimum counts and bacteria below a minimum log abundance.
Run this before running BRACoD because the algorithm does not perform well when there are many
low abundance bacteria that are only present in a few samples.
}
