% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRACoD.R
\name{install_bracod}
\alias{install_bracod}
\title{Install BRACoD in python}
\usage{
install_bracod(method = "auto", conda = "auto")
}
\description{
Uses pip to install the latest BRACoD release in python. You might need
to specify a python environment with either reticulate::use_virtualenv or
reticulate::use_condaenv.
}
