% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu-data.R
\docType{data}
\name{obesity}
\alias{obesity}
\title{Example microbiome data}
\format{
a DataFrame of 16S microbiome counts
}
\usage{
data(obesity)
}
\description{
This data is mouse stool microbiome data from a study of obesity.
}
\keyword{datasets}
