% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRACoD.R
\name{score}
\alias{score}
\title{Score the results of BRACoD}
\usage{
score(bugs_identified, bugs_actual)
}
\description{
This calculate the precision, recall and F1 of your BRACoD results if you know
the ground truth, ie. if this is simulated data.
}
